Remove unreachable code outside the calendar primitives
A reference sweep over every target turned up a handful of declarations
nothing calls. `HomeLogRow` is only built by its own preview — the log
section on Home inlines the same row layout itself — so the view goes
away. `MetricValue.intValue`, `ActivityPeriod.acronym` and the
`maxDuration`/`averageDuration` pair on a hang group have no callers at
all, not even in tests.

`Date.addingHour`/`addDay` and `Telemetry.export` are live, but only
from tests, so they move to where they are used: the two date helpers
join `Tests/Support`, which both test targets already depend on, and the
telemetry case mirror becomes a fileprivate helper in the one suite that
asserts through it. The tests that covered the moved date helpers are
dropped, since `adding(_:value:)` and `addingDay` stay covered by the
rest of the suite.

`HTTPWriteAck.saved` is deliberately kept: the field is never read, but
declaring it makes decoding reject an ack that omits it, so removing it
would loosen the wire contract rather than delete dead code.
