+++ Sources/Scout/Diagnostics/Telemetry/Telemetry.Export.swift
@@ -46,19 +46,4 @@ extension Telemetry {
             self = .timer(value)
         }
     }
-
-    var export: Export {
-        switch self {
-        case .counter:
-            .counter
-        case .floatingCounter:
-            .floatingCounter
-        case .meter:
-            .meter
-        case .recorder:
-            .recorder
-        case .timer:
-            .timer
-        }
-    }
 }
